Split NoConstructorArgumentsForInterfaceMock_2
