+++ Source/Moq.Analyzers.Test/NoConstructorArgumentsForInterfaceMockAnalyzerTests.ShouldPassIfRealMoqIsUsedWithValidParameters.verified.txt
